+++ cached_path/schemes/s3.py
@@ -10,7 +10,7 @@ import botocore
 from overrides import overrides
 
 from cached_path.common import _split_cloud_path
-from cached_path.protocols.cacher import Cacher
+from cached_path.schemes.cacher import Cacher
 
 
 class S3Cacher(Cacher):